% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BERTModule.R
\name{BERT.graphics.device}
\alias{BERT.graphics.device}
\title{Create a BERT graphics device.}
\usage{
BERT.graphics.device(name = "BERT-default", bgcolor = "white",
  width = 400, height = 400, pointsize = 14, cell = F)
}
\description{
\code{BERT.graphics.device} creates a graphics device that renders to 
a named Shape in an Excel workbook.
}
\details{
If this is called from an Excel spreadsheet cell, set \code{cell=TRUE} 
and it will use a name that's generated from the cell address, which 
should be unique in the workbook.  (That won't survive sheet renaming, 
but in that case just toss the old one and a new one will be generated 
on the next paint).

Size, background and pointsize arguments are ignored if the target 
named shape already exists.
}
